Return with_metaclass to the definition given in six
Specifically:

Do not hardcode __call__ and __init__ to be the versions from the
'type' metaclass.
ALL metaclasses in SymPy derive from type anyway, so those that want
and need these versions can have them anyway, and those that want to
override them can do so, too.
(Singleton will want to override __call__.)

Do not check for this_bases is None.
It cannot be None anyway, it can only be ().
(Actually not a single with_metaclass makes use of that "you can omit
'object'" featurem, so coding for that case is just potential bugs for
no gain.)

Use type.__new__(metaclass, ...) instead of metaclass(...).
I am not 100% sure about the consequences of this, but I'm simply going
with the wisdom of the six guys.
Might be necessary to avoid endless recursion in some cases.