Selection of integer-valued solutions
Quadratic equations with nonsquare discriminant are solved by
transforming to generalized Pell equations ``X**2 - D*Y**2 = N``.
The rational solutions of the two equations correspond to each other.
In addition, the transformation can be chosen such that all integer-
valued solutions of the original equation correspond to integer-valued
solutions of the Pell equation.
However, not all integer-valued solutions of the Pell equation
correspond to integer-valued solutions of the original equation,
because the rational coefficients of the inverse transformation are
not integers, in general.

The integer-valued solutions of the generalized Pell equation are as
follows. There is a finite number of fundamental solutions ``(X, Y)``,
and all solutions are obtained as coefficients of 1 and ``d = sqrt(D)``
in ``(X + d*Y)*(U + d*V)`` for all solutions of the Pell equation
``U**2 - D*V**2 = 1``. On the other hand, the solutions of the
Pell equation are given by the powers ``U_i + d*V_i = (U + d*V)**i``
of a single primitive solution ``(U, V)`` and their negatives.

The integer solutions of the original equation are then found with
the following steps.

    1. Find the fundamental solutions of the generalized Pell equation.

    2. Find the primitive solution ``(U, V)`` of Pell's equation.

    3. Find the least positive integer ``k`` such that ``(U_k, V_k)``
       is congruent to ``(1, 0)`` modulo the lcm ``L`` of the
       denominators of the coefficients of the inverse transformation.

    4. For each fundamental solution ``(X, Y)`` find the solutions
       ``(x_i, y_i)`` of the original equation corresponding to
       ``X_i + d*Y_i = (X + d*Y)*(U + d*V)**i for i in range(k)``.

    5. For each pair ``(x_i, y_i)`` of integer-valued solutions
       add all solutions corresponding to
       ``(X_i + d*Y_i)*(U_k + d*V_k)**t`` for all integers ``t``.