pretty print: use single-char square root symbol in simple cases
For single-char Symbols and for non-negative integers we use
the unicode square root symbol.  As this doesn't include a top
overline indicated what the root encompasses, we use this only
for simple unambiguous cases like:
````
>>> pprint(exp(3*sqrt(5*x*pi)))
 3⋅√5⋅√π⋅√x
ℯ
````
Update tests for this.

Perhaps slightly more controversial:
````
>>> pprint(sqrt(123))
√123
>>> pprint(sqrt(Symbol('C1')))
  ____
╲╱ C₁
````
Note---as in this last case---we do *not* simply use "√" for all
Symbols because "√C₁" looks ambiguous.  Add test for this.

An alternative would be to combine the "√" symbol with some
overlining, although at some cost in vertical space.  And the
overline does not connect to the "√" symbol.