Removing Sergey B Kirpichev from contributors
Sergey asked before to do this in case he will be blocked by SymPy
team on Github. After he was blocked, he reiterated this wish.

He also asked to explain here why he was blocked and provide links to
his posts that are offending. He was blocked because of his comments
in issues and PRs attacking people personally. The final post and
reason for ban was

https://github.com/sympy/sympy/issues/9263#issuecomment-91876155

    @debugger22, can you read your [tutorial](http://docs.sympy.org/dev/tutorial/gotchas.html#equals-signs) eventually?  Looks really odd for core team member...

    I have no idea how you "fixed the formatting" (@proy87, is that correct?), but now the bug description doesn't have problem at all.  The ``==`` is for structural equality in the SymPy, it wont do non-trivial simplifications.

    And why evalf label?

Some previous comments in no particular order are e.g.:

https://github.com/sympy/sympy/pull/9114#issuecomment-78131746
    (comment after which I warned Sergey)

    > But they should be separate, as it's otherwise confusing to users.

    For sure, references to missing documentation (e.g. to manualintegrate, or to ratint) can't confuse users...  @toolforger, is there any chance that you will read the code first?

https://github.com/sympy/sympy/pull/7356#issuecomment-39136345
https://github.com/sympy/sympy/pull/7295#issuecomment-38001055
https://github.com/sympy/sympy/pull/8860#discussion_r25420723
https://github.com/sympy/sympy/pull/9114#issuecomment-78194877
https://github.com/sympy/sympy/pull/9114#issuecomment-78474159
https://github.com/sympy/sympy/pull/9114#issuecomment-84004243
https://github.com/sympy/sympy/issues/7971#issuecomment-54419425
https://github.com/sympy/sympy/issues/7971#issuecomment-54511434
https://github.com/sympy/sympy/pull/7270#discussion_r10608484

I tried to resolve the issue and asked Sergey to please not attack
people, but he asked me to block him if I do not like his comments,
rather than change the way he interacts.

Sergey approved this message.